+++ src/fl_classifier/__init__.py
@@ -1,7 +1,7 @@
 # src/fl_classifier/__init__.py
 from .fl_classifier import main
 
-__version__ = "0.1.3"
+__version__ = "0.2.0"
 
 if __name__ == "__main__":
     main()